scaling progress
added third enemy. This is helping to break up the options when spawning

added weight to the spawning. so that the strongest enemy, the turret, will spawn less frequently

added a Timer Increase to the spawner as the difficulty increases
The rate of spawning will now decrease, so that when the game gets a little tougher, it's balanced by allowing a little breathing space