Boss scene in game
The boss scene is in game now.
there is an issue with the 2 Camera2D's ("C2D") being used
> in Scene Switcher ("SS")
> in Level01 ("L1")
I need the C2D in L1 so that i can do the screen zoom transition, but it prioritises the C2D in sceneswitcher. Disabling the C2D in SS will allow the transition to take place, but leaves the menu in a weird spot.

Still working to iron out this bug